migrate TestActivateVotingPeriod to simapp
